+++ apps-receipt-rewards/src/main/java/com/mytiki/apps_receipt_rewards/home/ui/HomeScreen.kt
@@ -1,84 +1,112 @@
 package com.mytiki.apps_receipt_rewards.home.ui
 
+import android.content.res.Configuration
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.core.Spring
+import androidx.compose.animation.core.SpringSpec
 import androidx.compose.animation.core.spring
+import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideInVertically
+import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.animation.slideOutVertically
 import androidx.compose.animation.togetherWith
+import androidx.compose.foundation.layout.requiredHeight
 import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.SheetValue
+import androidx.compose.material3.SheetState
 import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.ui.platform.LocalConfiguration
-import androidx.lifecycle.viewmodel.compose.viewModel
-import androidx.navigation.NavHostController
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.IntOffset
+import androidx.compose.ui.unit.dp
+import androidx.navigation.NavController
+import androidx.navigation.NavGraphBuilder
+import androidx.navigation.compose.composable
 import com.mytiki.apps_receipt_rewards.Rewards
 import com.mytiki.apps_receipt_rewards.account.AccountProvider
 import com.mytiki.apps_receipt_rewards.account.AccountType
-import com.mytiki.apps_receipt_rewards.home.HomeViewModel
-import com.mytiki.apps_receipt_rewards.ui.RewardsSharedViewModel
 import com.mytiki.apps_receipt_rewards.utils.components.BottomSheet
 import com.mytiki.apps_receipt_rewards.utils.navigation.RewardsRoute
 import kotlinx.coroutines.launch
 
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun HomeScreen(
-    navController: NavHostController,
-    onDismissBottomSheet: () -> Unit,
+@ExperimentalMaterial3Api
+class HomeScreen(
+    private val navGraphBuilder: NavGraphBuilder,
+    private val navController: NavController,
+    private val springSpec: SpringSpec<IntOffset>,
+    private val configuration: Configuration,
+    private val onDismiss: () -> Unit
 ) {
-    val configuration = LocalConfiguration.current
 
-    val sheetState = rememberModalBottomSheetState {
-        Rewards.showBottomSheet = it == SheetValue.Expanded
-        return@rememberModalBottomSheetState true
-    }
-    val scope = rememberCoroutineScope()
-
-    fun close() {
-        scope.launch { sheetState.hide() }.invokeOnCompletion {
-            onDismissBottomSheet()
+    fun route() {
+        navGraphBuilder.composable(
+            route = RewardsRoute.HomeScreen.name,
+            enterTransition = {
+                slideInHorizontally(
+                    initialOffsetX = { configuration.screenWidthDp / 2 },
+                    animationSpec = springSpec
+                )
+            },
+            exitTransition = {
+                slideOutHorizontally(
+                    targetOffsetX = { -configuration.screenWidthDp / 2 },
+                    animationSpec = springSpec
+                )
+            },
+            popEnterTransition = {
+                slideInHorizontally(
+                    initialOffsetX = { -configuration.screenWidthDp / 2 },
+                    animationSpec = springSpec
+                )
+            },
+            popExitTransition = {
+                slideOutHorizontally(
+                    targetOffsetX = { configuration.screenWidthDp / 2 },
+                    animationSpec = springSpec
+                )
+            }
+        ) {
+            val scope = rememberCoroutineScope()
+            val sheetState = rememberModalBottomSheetState(true)
+            BottomSheet(
+                sheetState = sheetState,
+                modifier = Modifier.requiredHeight(538.dp),
+                onDismiss = {
+                    scope.launch {
+                        sheetState.hide()
+                    }
+                }
+            ) {
+                AnimatedContent(
+                    targetState = sheetState.hasExpandedState,
+                    transitionSpec = {
+                        slideInVertically(
+                            animationSpec = spring(dampingRatio = Spring.DampingRatioNoBouncy),
+                        ) togetherWith
+                                slideOutVertically(
+                                    animationSpec = spring(dampingRatio = Spring.DampingRatioNoBouncy),
+                                )
+                    }, label = ""
+                ) { targetExpanded ->
+                    if (targetExpanded) {
+                        HomeExpanded { toAccount(it) }
+                    } else {
+                        HomeView( toAccount(it) ){
+                           scope.launch {
+                                sheetState.hide()
+                           }.invokeOnCompletion { onDismiss }
+                        }
+                    }
+                }
+            }
         }
     }
 
-    fun toAccount(accountProvider: AccountProvider) {
-        Rewards.currentProvider(accountProvider)
+    private fun toAccount(accountProvider: AccountProvider) {
+        Rewards.currentProvider = accountProvider
         if (accountProvider.accountType == AccountType.EMAIL) {
             navController.navigate(RewardsRoute.EmailScreen.name)
         } else {
             navController.navigate(RewardsRoute.RetailerScreen.name)
-        }
-    }
-
-    BottomSheet(
-        sheetState = sheetState,
-        onDismiss = {
-            Rewards.showBottomSheet = false
-            onDismissBottomSheet()
-        }
-    ) {
-        AnimatedContent(
-            targetState = homeViewModel.isExpanded.value,
-            transitionSpec = {
-                slideInVertically(
-                    animationSpec = spring(dampingRatio = Spring.DampingRatioNoBouncy),
-                ) togetherWith
-                        slideOutVertically(
-                            animationSpec = spring(dampingRatio = Spring.DampingRatioNoBouncy),
-                        )
-            }, label = ""
-        ) { targetExpanded ->
-            if (targetExpanded) {
-                HomeExpanded(homeViewModel) { toAccount(it) }
-            } else {
-                HomePartiallyExpanded(
-                    homeViewModel,
-                    { navController.navigate(RewardsRoute.MoreScreen.name) },
-                    { toAccount(it) }) { close() }
-
-            }
         }
     }
 }
