+++ apps-receipt-rewards/src/main/java/com/mytiki/apps_receipt_rewards/terms/OfferTermsScreen.kt
@@ -1,36 +1,44 @@
-package com.mytiki.apps_receipt_rewards.offer.ui
+package com.mytiki.apps_receipt_rewards.terms
 
 import android.content.res.Configuration
 import androidx.compose.animation.core.SpringSpec
 import androidx.compose.animation.slideInHorizontally
 import androidx.compose.animation.slideOutHorizontally
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.SheetState
-import androidx.compose.material3.rememberModalBottomSheetState
-import androidx.compose.material3.rememberStandardBottomSheetState
-import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Divider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Scaffold
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
-import com.mytiki.apps_receipt_rewards.utils.components.BottomSheet
-import kotlinx.coroutines.launch
 import androidx.navigation.compose.composable
+import com.mytiki.apps_receipt_rewards.Rewards
+import com.mytiki.apps_receipt_rewards.utils.components.Header
+import com.mytiki.apps_receipt_rewards.utils.components.MainButton
 import com.mytiki.apps_receipt_rewards.utils.navigation.RewardsRoute
 
-@OptIn(ExperimentalMaterial3Api::class)
-class OfferScreen(
+
+class OfferTermsScreen(
     private val navGraphBuilder: NavGraphBuilder,
     private val navController: NavController,
     private val springSpec: SpringSpec<IntOffset>,
     private val configuration: Configuration,
-    private val onDismiss: () -> Unit
 ) {
-    fun route(){
+    fun route() {
         navGraphBuilder.composable(
-            route = RewardsRoute.OfferScreen.name,
+            route = RewardsRoute.TermsScreen.name,
             enterTransition = {
                 slideInHorizontally(
                     initialOffsetX = { configuration.screenWidthDp / 2 },
@@ -56,17 +64,7 @@ class OfferScreen(
                 )
             }
         ) {
-            val scope = rememberCoroutineScope()
-            val sheetState = rememberModalBottomSheetState(true)
-            BottomSheet(
-                sheetState = sheetState,
-                modifier = Modifier.requiredHeight(538.dp),
-                onDismiss =  {
-                    scope.launch {
-                        sheetState.hide()
-                    }.invokeOnCompletion { onDismiss }
-                }
-            ) { OfferContent(navController) }
+            OfferTerms(navController)
         }
     }
 }
